refactored code. ros2 ws builds
