+++ base-station-ros2/src/base_station_coms/include/base_station_coms/seatrac_enums.hpp
