+++ base-station-ros2/src/base_station_coms/src/base_station_coms.cpp
@@ -5,15 +5,17 @@
 #include "seatrac_interfaces/msg/modem_rec.hpp"
 #include "seatrac_interfaces/msg/modem_send.hpp"
 
-#include "cougars_coms/coms_protocol.hpp"
+#include "base_station_coms/coms_protocol.hpp"
+#include "base_station_coms/seatrac_enums.hpp"
 
 #include <iostream>
 #include <chrono>
 #include <memory>
 
 using namespace std::literals::chrono_literals;
-using std::placeholders::_1;
 using namespace cougars_coms;
+using namespace narval::seatrac;
+using std::placeholders::_1;
 
 class ComsNode : public rclcpp::Node {
 public:
@@ -29,9 +31,6 @@ public:
         );
         this->modem_publisher_ = this->create_publisher<seatrac_interfaces::msg::ModemSend>("modem_send", 10);
 
-        this->thruster_client_ = this->create_client<std_srvs::srv::SetBool>(
-            "arm_thruster"
-        );
     }
 
     void listen_to_modem(seatrac_interfaces::msg::ModemRec msg) {
@@ -39,47 +38,14 @@ public:
         switch(id) {
             default: break;
             case EMPTY: break;
-            case EMERGENCY_KILL: {
-                kill_thruster();
-            } break;
         }
-    }
-
-    void kill_thruster() {
-        auto request = std::make_shared<std_srvs::srv::SetBool::Request>();
-        request->data = false;
-        while (!this->thruster_client_->wait_for_service(1s)) {
-            if (!rclcpp::ok()) {
-                RCLCPP_ERROR(rclcpp::get_logger("rclcpp"), 
-                    "Interrupted while waiting for the arm_thruster service. Exiting.");
-            }
-            RCLCPP_INFO(rclcpp::get_logger("rclcpp"), "arm_thruster service not available, waiting again...");
-        }
-
-        auto result_future = this->thruster_client_->async_send_request(request,
-            [this](rclcpp::Client<std_srvs::srv::SetBool>::SharedFuture response_future) {
-                try {
-                    auto response = response_future.get();
-                    if (response->success) {
-                        RCLCPP_INFO(rclcpp::get_logger("rclcpp"), "Thruster has been deactivated.");
-                    } else {
-                        RCLCPP_ERROR(rclcpp::get_logger("rclcpp"), "Failed to deactivate thruster.");
-                    }
-                    ConfirmEmergencyKill msg;
-                    msg.success = response->success;
-                    this->send_acoustic_message(base_station_beacon_id_, 1, (uint8_t*)&msg);
-                } catch (const std::exception &e) {
-                    RCLCPP_ERROR(this->get_logger(), "Error while trying to deactivate thruster: %s", e.what());
-                }
-            }
-        );
     }
 
     void send_acoustic_message(int target_id, int message_len, uint8_t* message) {
         auto request = seatrac_interfaces::msg::ModemSend();
-        request.msg_id = 0x60; //CID_DAT_SEND
+        request.msg_id = CID_DAT_SEND;
         request.dest_id = (uint8_t)target_id;
-        request.msg_type = 0x0; //MSG_OWAY, data sent one way without response or position data
+        request.msg_type = MSG_OWAY;
         request.packet_len = (uint8_t)std::min(message_len, 31);
         std::memcpy(&request.packet_data, message, request.packet_len);
         
@@ -91,7 +57,6 @@ private:
 
     rclcpp::Subscription<seatrac_interfaces::msg::ModemRec>::SharedPtr modem_subscriber_;
     rclcpp::Publisher<seatrac_interfaces::msg::ModemSend>::SharedPtr modem_publisher_;
-    rclcpp::Client<std_srvs::srv::SetBool>::SharedPtr thruster_client_;
 
     int base_station_beacon_id_;
